requirements.txt
scrapling>=0.2.9
pydoll-python>=1.0.0
aiohttp>=3.9.0
lxml>=4.9.0
python-dotenv>=1.0.0
